kvclient: split legacy rangefeed code out
Move the legacy rangefeed code to its own file to make it easy to delete
once it is no longer needed. This also helps with partial partition
handling for changefeeds as it won't include legacy code.

Release note: None
